feat: update JS files to TS
